rurminer: add support for sub-projects in runner
If you run `cargo test` in any of the subprojects, it fails, because
it runs with working-directory set to project directory - which is not
the rurminer root directory where the runner script
`scripts/runner.py` resides.

This commit adds path-like functionality to runner (to find Test.toml)
and locates runner script in parent directory if neccessary.