rurminer: Fix failing test of irq timeout
When wait-irq-timeout test was run right after proper mining, there
was a stale irq notification left in uio that wasn't read (because the
solution that triggered the interrupt arrived _after_ rurminer
software exited) and caused the test to fail with a irq that wasn't
expected.

Now we flush queues during fifo initialization so this doesn't happen
anymore.