rurminer: Initialize fifos in hchain-ctl
This fixes bug when the fifos were lazily initialized some time after
initialization, which is not OK since the reset the queues.

Now we allocate them at the time of initialization and the rx/tx tasks
just take them from the main hchain structure.

This also prepares ground for FIFO register window separation.