bosminer: s9: Change int. test into unit test
`s9_test` needs access to internals of S9 backend to do meaningful
testing. This internals also shouldn't be published as API to outside
crates (ie. `tests`).

It was agreed that tests that require specific hardware can be
implemented as unit tests (until now they had to be integration test)
if they are located in hardware-specific module.