bosminer: s9: Make type of voltage backend fixed
The HashChain voltage backend doesn't in reality change, but lugging
around `VBackend` and all the bounds it requires at every place that
uses HashChain is really tiresome.

In future: either make it a dyn trait, or do a better analysis at
which time (compile time? run time?) is the voltage backend determined
and implement it accordingly.